+++ monorepo/apps/blog/src/app/blog/_components/PostList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./PostList";
