+++ monorepo/apps/blog/src/app/blog/_components/TagList/TagList.tsx
@@ -1,0 +1,30 @@
+"use client";
+
+import Tag from "@/components/Tag";
+import useGetTags from "@/apis/useGetTags";
+import useTagContext from "../../_hook/useTagContext";
+
+const TagList = () => {
+  const { tagsData } = useGetTags();
+
+  const { selectedTag, setSelectedTag } = useTagContext();
+
+  return (
+    <div className="flex justify-center m-[20px] flex-wrap">
+      {tagsData.map((tag) => {
+        const isSelected = tag === selectedTag;
+
+        return (
+          <Tag
+            key={tag}
+            tag={tag}
+            isSelected={isSelected}
+            onClick={() => setSelectedTag(tag)}
+          />
+        );
+      })}
+    </div>
+  );
+};
+
+export default TagList;
