+++ monorepo/apps/blog/src/app/blog/_components/TagList/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TagList";
