+++ monorepo/apps/blog/src/app/blog/layout.tsx
@@ -1,0 +1,19 @@
+import HomeButtonWrapper from "@/components/HomeButtonWrapper";
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "게시글 목록",
+};
+
+export default function BlogLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <HomeButtonWrapper />
+      {children}
+    </>
+  );
+}
